+++ lib/src/utils/typedefs.dart
@@ -1,0 +1,5 @@
+import 'dart:async';
+
+typedef Reducer<T, U> = FutureOr<T> Function(T currentState, U action);
+
+typedef Dispatch<T, U> = FutureOr<T> Function(U action);