+++ lib/src/widgets/widget.dart
@@ -1,26 +1,23 @@
 // Observer
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:o/src/core/core.dart';
 
-class Observer<T> extends StatelessWidget {
-  final Observable<T> observable;
-  final Widget Function(BuildContext context, Observable<T> obs, T value)
-      builder;
+class Observer<T, U> extends StatelessWidget {
+  final Observable<T, U> observable;
+  final Widget Function(BuildContext context, T value) builder;
 
   /// An [Observable] listener.
   ///
   /// *Args*
   /// >-  `observable`
-  const Observer({required this.observable, required this.builder, Key? key})
-      : super(key: key);
+  const Observer({required this.observable, required this.builder, super.key});
 
   @override
-  Widget build(context) {
-    return StreamBuilder<T>(
-      stream: observable.stream,
-      builder: (context, value) {
-        return builder(context, observable, observable.value);
-      },
-    );
-  }
+  Widget build(context) => StreamBuilder<T>(
+        initialData: observable.value,
+        stream: observable.stream,
+        builder: (context, value) {
+          return builder(context, observable.value);
+        },
+      );
 }